requirements.txt
langchain == 0.1.4
langchain-community == 0.0.16
langchain-core == 0.1.17
ctransformers == 0.2.27
beautifulsoup4 == 4.12.3
duckduckgo_search == 4.4
faiss-cpu == 1.7.4
sentence-transformers
streamlit
chromadb
pypdf
huggingface_hub
python-dotenv